Amend plutus.py to use new get_balance function.
Plutus now has three balance functions that return balances in Satoshis,
which should help resolve issues such as HTTP 429 "slow your mojo".

When exceptions are caught, these are assumed to be the server blocking
us for some reason, so the global variables for each balance functions are
set to True.

When they are all set to False, then Plutus fails to deliver balances.

Working on a fix for this in case a server sends a Retry-After HTTP header,
so we can keep an eye on a timer and then resume using the balance API
call.

Occasionally weird errors happen, so an exception is caught and it's
simply ignored without further ado if the server returns strange JSON
that can't be parsed, or returns encoding which can't be detected.